feat: move `FetchError` to `src/constants/error.ts`
