+++ src/app/(normal)/pages/Profile.tsx
@@ -15,9 +15,9 @@ import { ProfileSourceTabs } from '@/components/Profile/ProfileSourceTabs.js';
 import { Title } from '@/components/Profile/Title.js';
 import { SuspendedAccountFallback } from '@/components/SuspendedAccountFallback.js';
 import { PageRoute, Source } from '@/constants/enum.js';
+import { FetchError } from '@/constants/error.js';
 import { SITE_NAME } from '@/constants/index.js';
 import { createPageTitle } from '@/helpers/createPageTitle.js';
-import { FetchError } from '@/helpers/fetch.js';
 import { isRoutePathname } from '@/helpers/isRoutePathname.js';
 import { narrowToSocialSource } from '@/helpers/narrowSource.js';
 import { useUpdateCurrentVisitingProfile } from '@/hooks/useCurrentVisitingProfile.js';
